backend/requirements.txt
fastapi>=0.110
uvicorn>=0.27
python-telegram-bot>=21.0
fastf1>=3.8
matplotlib>=3.8
Pillow>=10.0
httpx>=0.27
redis>=5.0
python-dotenv>=1.0
pandas>=2.0
pydantic>=2.0
pydantic-settings>=2.0
pyyaml>=6.0
